# src/generated/server/model/apiResponse.ts
/**
 * Color API
 * API for managing user\'s colors
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiResponse { 
    /**
     * HTTP status code
     */
    statusCode: number;
    /**
     * Response data payload
     */
    data?: any | null;
    /**
     * Optional message (primarily for errors)
     */
    message?: string;
}