# src/generated/server/model/colorRecord.ts
/**
 * Color API
 * API for managing user\'s colors
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ColorRecord { 
    /**
     * Primary key (firstName)
     */
    pk: string;
    /**
     * Array of colors submitted by the user
     */
    colors: Array<string>;
    /**
     * Timestamp of the submission
     */
    timestamp: string;
}