# src/generated/server/model/colorSubmission.ts
/**
 * Color API
 * API for managing user\'s colors
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ColorSubmission { 
    /**
     * User\'s first name (used as primary key)
     */
    firstName: string;
    /**
     * User\'s color
     */
    color: string;
}