# src/generated/server/model/errorResponse.ts
/**
 * Color API
 * API for managing user\'s colors
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ErrorResponse { 
    /**
     * HTTP status code
     */
    statusCode: number;
    data?: any | null;
    /**
     * Optional message (primarily for errors)
     */
    message: string;
}